requirements.txt
pymssql==2.2.7

streamlit

langchain

langchain-community

charset-normalizer==2.0.9

openai

pyyaml

python-dotenv

sqlalchemy

streamlit-chat

scikit-learn

selenium

beautifulsoup4

ipykernel








